Lagt til kopier saksnr-knapp + flytta oppgaverad ut i eigen komponent
