+++ packages/los-saksbehandler/src/behandlingskoer/oppgaveTabell/oppgaveRad.module.css
@@ -1,0 +1,11 @@
+.isUnderBehandling {
+  background-color: var(--a-orange-100);
+}
+
+.image {
+  color: var(--a-blue-600);
+  height: 30px;
+  margin-bottom: -5px;
+  margin-top: -5px;
+  width: 30px;
+}
