+++ packages/los-saksbehandler/src/behandlingskoer/oppgaveTabell/oppgaverTabell.module.css
@@ -1,15 +1,3 @@
-.isUnderBehandling {
-  background-color: var(--a-orange-100);
-}
-
-.image {
-  color: var(--a-blue-600);
-  height: 30px;
-  margin-bottom: -5px;
-  margin-top: -5px;
-  width: 30px;
-}
-
 .grayout {
   color: var(--a-gray-500);
 }
